first draft of permissions
