+++ packages/app/src/app/pages/Dashboard/Content/routes/Settings/TeamSettings/index.tsx
@@ -1,0 +1,1 @@
+export { TeamSettings } from './TeamSettings';
